Tests/LibWeb/Text/input/css/style-invalidation/has-interaction-pseudo-insertion-invalidation.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="../../include.js"></script>
<style>
    /* :has() with an interaction pseudo-class argument, anchored in
       non-subject position at an ancestor of the streaming container. Plain
       insertions can never change whether interaction pseudo-classes match,
       so appends must not restyle the wrapper subtree. */
    #wrapper:has(:focus-visible) span {
        color: rgb(9, 9, 9);
    }

    #wrapper:has(:focus) span {
        color: rgb(7, 7, 7);
    }

    span {
        color: rgb(1, 1, 1);
    }
</style>
<div id="wrapper">
    <div id="big"></div>
    <div id="stream"></div>
    <input id="field">
</div>
<script>
    test(() => {
        const big = document.getElementById("big");
        for (let i = 0; i < 500; ++i) {
            const leaf = document.createElement("span");
            leaf.textContent = `leaf ${i}`;
            big.appendChild(leaf);
        }
        const stream = document.getElementById("stream");
        internals.updateStyle();
        internals.resetStyleInvalidationCounters();

        for (let i = 0; i < 10; ++i) {
            const p = document.createElement("p");
            p.textContent = `chunk ${i}`;
            stream.appendChild(p);
            internals.updateStyle();
        }

        const counters = internals.getStyleInvalidationCounters();
        if (counters.elementStyleRecomputations < 100)
            println("PASS: appends near interaction :has() stayed targeted");
        else
            println(`FAIL: appends near interaction :has() recomputed ${counters.elementStyleRecomputations} elements`);

        // Focus changes and removal of a focused element must still invalidate
        // the :has(:focus) anchor.
        const someLeaf = big.firstElementChild;
        const field = document.getElementById("field");
        field.focus();
        println(`focused: ${getComputedStyle(someLeaf).color}`);
        field.remove();
        println(`focused element removed: ${getComputedStyle(someLeaf).color}`);
    });
</script>
