Tests/LibWeb/Text/expected/css/style-invalidation/has-interaction-pseudo-insertion-invalidation.txt

FAIL: appends near interaction :has() recomputed 5020 elements
focused: rgb(7, 7, 7)
focused element removed: rgb(1, 1, 1)
